refactor(compat): move useSyncExternalStore into render
