+++ mangle.json
@@ -75,7 +75,6 @@
       "$_unmount": "__u",
       "$_owner": "__o",
       "$_skipEffects": "__s",
-      "$_hydrationRoot": "__y",
       "$_forwarded": "__f",
       "$_isSuspended": "__i",
       "$_bits": "__g"
